aenea/requirements.txt
beautifulsoup4==4.4.1
certifi
future==0.15.2
python-telegram-bot==5.0.0
requests
urllib3
pyasn1
pyasn1_modules